requirements.txt
streamlit
tensorflow
numpy
pillow
fpdf
gdown
